whisper-gui/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "my_package"
authors = [
    {name = "Kevin James", email = "krjii@indyhustles.com"},
]
description = "My package description"
readme = "README.rst"
requires-python = ">=3.11"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "openai-whisper",
    "setuptools-rust",
    "pyside6",
    "TextEdit",
    "ffmpeg"
]
dynamic = ["version"]

[project.optional-dependencies]
pdf = ["ReportLab>=1.2", "RXP"]
rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
my-script = "my_package.module:function"

# ... other project metadata fields as listed in:
#     https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
# Using open-ai whisper -> https://github.com/openai/whisper?tab=readme-ov-file